requirements.txt
flask
gunicorn
flask_sqlalchemy
flask_login
flask_bcrypt
psycopg2-binary
python-dotenv